moving model-specific sdxl code to sdxl_pipeline.py, renaming sdxl_pipeline_benchmark to sdxl_pipeline
